timer - measurement period ~5s

+++ app/SENSORS/lm35.h
@@ -8,7 +8,7 @@ namespace SENSORS{
         public:
         ADC::PinADC1 adc;
         void setADC(ADC::PinADC1 _adc);
-        double readTemperature();
+        float readTemperature();
     };   
 }
 #endif
